feat(migration): add migration & re-structure dal
